Move account payment order and line fields to l10n_br_account_payment_order
