+++ l10n_br_account_payment_order/views/account_move_line.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
+
     <record id="cnab_view_move_line_form_inherit" model="ir.ui.view">
         <field name="name">cnab.account.move.line.form.inherit</field>
         <field name="model">account.move.line</field>
@@ -20,13 +21,11 @@
         </field>
     </record>
 
-    <record id="l10n_br_payment_boleto_view_payments_tree"
-            model="ir.ui.view">
+    <record id="l10n_br_payment_boleto_view_payments_tree" model="ir.ui.view">
         <field name="name">Payments</field>
         <field name="model">account.move.line</field>
         <field name="field_parent">partner_id</field>
-        <field name="inherit_id"
-               ref="account_due_list.view_payments_tree"/>
+        <field name="inherit_id" ref="account_due_list.view_payments_tree"/>
         <field name="arch" type="xml">
             <field name="date_maturity" position="before">
                 <field name="date_payment_created" readonly="1"/>
