+++ l10n_br_account_payment_order/views/account_payment_order.xml
@@ -14,6 +14,47 @@
             <field name="payment_line_ids" position="replace">
                 <field name="payment_line_ids" context="{'default_payment_type': payment_type}"/>
             </field>
+            <div class="oe_title" position="before">
+                <div class="alert alert-info" role="alert"
+                     style="margin-bottom:0px;"
+                     attrs="{'invisible': [('bank_line_error_ids','=',[])]}">
+                    <bold>
+                        <a class="alert-link">
+                           Arquivo gerado com problemas...
+                        </a>
+                    </bold>
+                    Verifique-os na aba "Problemas"
+                </div>
+            </div>
+            <field name="bank_line_count" position="after">
+                <field name="service_type"/>
+                <field name="release_form"/>
+                <field name="code_convetion"/>
+            </field>
+            <field name="company_id" position="after">
+                <field name="indicative_form_payment"/>
+                <field name="movement_instruction_code"/>
+                <field name="movement_type"/>
+            </field>
+            <field name="payment_line_ids" position="attributes">
+                <attribute name="context">{'default_payment_type': payment_type, 'tree_view_ref': 'l10n_br_account_payment_cobranca.l10n_br_cobranca_payment_line_tree', 'form_view_ref': 'l10n_br_account_payment_cobranca.l10n_br_cobranca_payment_line_form'}</attribute>
+            </field>
+            <page name="bank-payment-lines" position="after">
+                <page name="bank-error-lines" string="Problemas"
+                      attrs="{'invisible': [('bank_line_error_ids','=',[])]}">
+                    <field name="bank_line_error_ids"
+                           context="{'default_payment_type': payment_type}">
+                        <tree>
+                            <field name="name"/>
+                            <field name="partner_id"/>
+                            <field name="own_number"/>
+                            <field name="document_number"/>
+                            <field name="company_title_identification"/>
+                            <field name="export_error_message"/>
+                        </tree>
+                    </field>
+                </page>
+            </page>
         </field>
     </record>
